Month small refactor
