+++ src/reducers/index.js
@@ -18,15 +18,45 @@ const initialState = {
         },
         {
           id: 3,
-          nameOfDay: 'Śr',
+          nameOfDay: 'ŚR',
           hours: 9,
           holiday: false,
         },
         {
           id: 4,
+          nameOfDay: 'CZ',
+          hours: 0,
+          holiday: false,
+        },
+        {
+          id: 5,
+          nameOfDay: 'PT',
+          hours: 0,
+          holiday: false,
+        },
+        {
+          id: 6,
+          nameOfDay: 'SO',
+          hours: 0,
+          holiday: false,
+        },
+        {
+          id: 7,
           nameOfDay: 'ND',
           hours: 0,
           holiday: true,
+        },
+        {
+          id: 8,
+          nameOfDay: 'PN',
+          hours: 0,
+          holiday: false,
+        },
+        {
+          id: 9,
+          nameOfDay: 'WT',
+          hours: 0,
+          holiday: false,
         },
       ],
     },
